Repository and its implementation were added for popular movies

+++ app/src/main/java/com/challenge/movies/popular/data/PopularMoviesResponse.kt
@@ -1,8 +1,8 @@
 package com.challenge.movies.popular.data
 
 data class PopularMoviesResponse(
-    val page: Long? = null,
+    val page: Int? = null,
     val results: List<PopularMovie>? = null,
-    val totalPages: Long? = null,
-    val totalResults: Long? = null
+    val totalPages: Int? = null,
+    val totalResults: Int? = null
 )
